Compensate to bug in Phishing.Database
Do to a bug in Phishing.Database we are not able to do full search in the active files. For that reason we are now importing the `ALL-phishing-links.txt` and strips it down to domain only list in `data/phishing_database/`

Related issues:
- https://github.com/mitchellkrogza/Phishing.Database/issues/840
- https://github.com/mitchellkrogza/Phishing.Database/issues/881
- https://github.com/mitchellkrogza/phishing/pull/381#issuecomment-2119335770
- https://github.com/mitchellkrogza/phishing/pull/396
- https://github.com/mitchellkrogza/phishing/pull/407
- https://github.com/mitchellkrogza/phishing/issues/395
- https://github.com/mypdns/matrix/issues/624
- https://github.com/blocklistproject/Lists/pull/1252
- https://github.com/mitchellkrogza/Phishing.Database/issues/840
- https://github.com/mitchellkrogza/Phishing.Database/issues/722

Trying to use @main for the php installer and using php version 8.4

Added `libdomain-publicsuffix-perl` to the dependencies.sh script as it is required by perl in import.sh. It turns out Perl just anoyingly does it again... :smirk: